Move testing fakes to their own namespace
