+++ Tvl.VisualStudio.MouseFastScroll.UnitTests/Fakes/FakeViewScroller.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Tunnel Vision Laboratories, LLC. All Rights Reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE.txt in the project root for license information.
 
-namespace Tvl.VisualStudio.MouseFastScroll.UnitTests
+namespace Tvl.VisualStudio.MouseFastScroll.UnitTests.Fakes
 {
     using System;
     using Microsoft.VisualStudio.Text;
